requirements.txt
# Data Ingestion
yfinance>=0.2.36
requests>=2.31.0
beautifulsoup4>=4.12.3
feedparser>=6.0.11

# Snowflake
snowflake-connector-python>=3.6.0
snowflake-sqlalchemy>=1.5.0

# Kafka
confluent-kafka>=2.3.0
kafka-python>=2.0.2

# dbt
dbt-snowflake>=1.7.0

# Unstructured Data Processing
textblob>=0.18.0
nltk>=3.8.1

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
schedule>=1.2.0
loguru>=0.7.0
